+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+<!-- next-header -->
+## [Unreleased] - ReleaseDate
 
 ### Added
 - Async support based on `embedded-hal-async` 1.0.
@@ -46,6 +47,8 @@ This crate is now functionally complete.
 This is the initial release to crates.io. All changes will be documented in
 this CHANGELOG.
 
+
+<!-- next-url -->
 [Unreleased]: https://github.com/eldruin/tcs3472-rs/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/eldruin/tcs3472-rs/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/eldruin/tcs3472-rs/compare/v0.1.0...v0.1.1